requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-allauth==0.61.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
langchain==0.1.9
openai==1.12.0
Pillow==10.2.0
python-jose==3.3.0
gunicorn==21.2.0
whitenoise==6.6.0
djangorestframework-simplejwt==5.3.1
dj-rest-auth==5.0.2
dj-rest-auth[with_social]==5.0.2 